Drop package initialization (#2)
This does not work correctly as `conda-build` strips out the `.git`
directory from `METADATA`. There are no clean fixes for this issue. As
we already try to initialize the package directory when Julia starts,
this will be handled anyways on the user's machine. In the end, this may
be best to avoid having a potentially out-of-date index in the package.