Remove wrapper script (#11) (#12)
* Removed wrapper script

* Bumped build number

* Use juliarc ( dfornika#1 )

* Add a custom juliarc.jl

Needed to ensure packages and other relevant information to the Julia
install are tracked in the conda environment.

* Install custom juliarc

Make sure that that our custom juliarc is available in the standard
location.

* Create the package directory

Make sure that Julia initializes our package directory and includes the
standard packages in it.

* Bump build number to 1

Needed to repackage after removal of the wrapper script, inclusion of
juliarc, and inclusion of the prepopulated package directory.

* Removed JULIA_HISTORY setting

* Drop package initialization ( dfornika#2 )

This does not work correctly as `conda-build` strips out the `.git`
directory from `METADATA`. There are no clean fixes for this issue. As
we already try to initialize the package directory when Julia starts,
this will be handled anyways on the user's machine. In the end, this may
be best to avoid having a potentially out-of-date index in the package.

* Append juliarc.jl instead of replace

* Append juliarc.jl the right way

* Added post-link script for Pkg.init()

* Removed Pkg.init() from juliarc.jl

* Correct setting Base.LOAD_CACHE_PATH

* Correct setting Base.LOAD_CACHE_PATH again

(cherry picked from commit 19d96fa559582c8162f92896a3047c268a397c58)