Ignore prefix in julia executables
Based on the warnings we are getting like the sample one below, it seems
like `conda-build`'s prefix fixing behavior may be messing up the
executables' behavior after the fact. To fix this, we tell `conda-build`
to not fix the prefix in these binaries (even if it may be wrong).

"warning: parsing line table prologue at 0x0000e1c8 should have ended at
0x0000e52f but it ended at 0x0000e284"