Require openlibm
This would be built internally by Julia as well. However switching to an
openlibm dependency is better principled from a package management
standard. Hence why we add it here

As we don't really know openlibm's compatibility, pin it exactly to the
version that Julia would have required for this release anyways. This
way we are guaranteed to match Julia's build even though we are
separating out this dependency. Once we know more about openlibm's
compatibility, we can look into loosening this pinning a bit.