Require openspecfun
This is a dependency of Julia's that is normally built internally.
However it can be built externally and used in the Julia build. In order
to better express Julia's requirements, we break this out and make it an
external dependency of Julia.

For now we pin openspecfun exactly as we are unsure of its compatibility
between versions. Plus we want to make sure it matches what Julia wants
it to be. We can revisit how to pin openspecfun once we know more about
its compatibility.