Require openspecfun (#26)
* Move BLAS and LAPACK flags before System flags

A minor stylistic change that makes it easier to keep all of the
`USE_SYSTEM_*` flags together and extend them without burying them in
the `USE_SYSTEM_*` flags.

* Require openspecfun

This is a dependency of Julia's that is normally built internally.
However it can be built externally and used in the Julia build. In order
to better express Julia's requirements, we break this out and make it an
external dependency of Julia.

For now we pin openspecfun exactly as we are unsure of its compatibility
between versions. Plus we want to make sure it matches what Julia wants
it to be. We can revisit how to pin openspecfun once we know more about
its compatibility.

* Drop gcc and libgcc

Neither of these are needed any more as all the compiled code is now C
or C++. Any portions that use Fortran are now broken out as
dependencies. So the regular system toolchain should be sufficient for
building Julia.

* Bump build number to 4

As this now has a new dependency, openspecfun, and no longer uses the
gcc package, we need to do a rebuild to get a package built with these
changed dependencies.