Specify path to our julia in link scripts
Instead of allowing an arbitrary julia on the path to be picked up in
the link scripts, ensure that we provide the full path to the julia
executable in our environment. This lesson has already been learned with
Jupyter packages that use link scripts. So we merely try to reuse that
knowledge here.