Redirect stderr/stdout from link scripts
With link scripts, anything written to stdout or stderr ends up getting
suppressed by `conda`. This actually is a smart thing for `conda` to do
in order to avoid having its output messed up by the link scripts or
worse having something fail because of it. To address the need for
output from these link scripts, `conda` provides an option to write to
`.messages.txt` in the environment `$PREFIX` to send this information to
the user. So we take advantage of this resource to share info about
initializing Julia's package index. This again borrows from the same
strategy already used in Jupyter link scripts of append to
`.messages.txt` should we need to write more.