Fix link/unlink step (#29)
* Cleanup Julia packages in unlink step

As part of Julia's removal, make sure to remove all packages that were
installed into the environment in the pre-unlink step.

* Specify path to our julia in link scripts

Instead of allowing an arbitrary julia on the path to be picked up in
the link scripts, ensure that we provide the full path to the julia
executable in our environment. This lesson has already been learned with
Jupyter packages that use link scripts. So we merely try to reuse that
knowledge here.

* Redirect stderr/stdout from link scripts

With link scripts, anything written to stdout or stderr ends up getting
suppressed by `conda`. This actually is a smart thing for `conda` to do
in order to avoid having its output messed up by the link scripts or
worse having something fail because of it. To address the need for
output from these link scripts, `conda` provides an option to write to
`.messages.txt` in the environment `$PREFIX` to send this information to
the user. So we take advantage of this resource to share info about
initializing Julia's package index. This again borrows from the same
strategy already used in Jupyter link scripts of append to
`.messages.txt` should we need to write more.

* Bump build number to 5

Need a rebuild to include the updated post-link and new pre-unlink
scripts in the new `julia` packages.

(cherry picked from commit d3b26341b2bc3e400b87e031a8277d836aec6acc)