Use SHLIB_EXT (#30)
Previously we were hard coding the library extensions for the BLAS and
LAPACK libraries with `.so` (the Linux shared object extension). However
this extension won't work on macOS where `.dylib` (the macOS dynamic
library extension) is used. In order to specify these libraries
correctly on both Linux and macOS, we switch to using `SHLIB_EXT`, which
`conda-build` sets correctly according to whether we are on Linux or
macOS. That way the correct library extension is always specified
regardless of what platform we are on.