Use patchelf on Linux only (#31)
This doesn't matter much while we are only building for Linux. However
once we start building for macOS or even Windows, we will want to ensure
that `patchelf` is only pulled on Linux. Hence we add this selector.
Though this is really a no-op as far as the Linux build is concerned.