Skip doc build (#32)
* Add a hack to suppress building the docs

There doesn't seem to be a formal way to disable building Julia's docs.
As a hack, we overwrite the Sphinx Makefile in the `docs` directory with
one that provides the desired `html` target, but does nothing more than
create the expected output directory with no content. By doing this, we
are able to skip building the docs, but ensure it doesn't mess up the
install step. This should cutdown the build time some and cutdown the
size of the `conda` package.

* Bump build number to 6

Rebuild package to exclude the HTML docs.

(cherry picked from commit c51a34c3a06e981249daa39fcbc6901385fd1bf6)