Shorten tests run on macOS
Travis CI is pretty restrictive on the time limit of each build. As
such we can only run a few tests as opposed to the full test suite at
the time. So we run the core test suite and backtrace as the latter
picked up some subtle build issues with LLVM. This should hopefully fit
in Travis CI's build time limit.