Require libosxunwind on macOS
As there some to be some build issue with parallel builds on macOS where
`libosxunwind` does not always get built, we have packaged
`libosxunwind` separately and now add it as a dependency of `julia`.
This should avoid the improper dependency resolution in the `make` build
currently by simply bypassing it. Also may speed up the build of `julia`
on macOS a little bit.